+++ tests/Utils/Models/User.php
@@ -30,6 +30,7 @@ use Illuminate\Foundation\Auth\User as Authenticatable;
  * @property string|null $person_type
  *
  * Relations
+ * @property-read \Illuminate\Database\Eloquent\Collection<\Tests\Utils\Models\AlternateConnection> $alternateConnections
  * @property-read \Tests\Utils\Models\Company|null $company
  * @property-read \Tests\Utils\Models\Image|null $image
  * @property-read \Illuminate\Database\Eloquent\Collection<\Tests\Utils\Models\Post> $posts
@@ -40,6 +41,11 @@ use Illuminate\Foundation\Auth\User as Authenticatable;
  */
 class User extends Authenticatable
 {
+    public function alternateConnections(): HasMany
+    {
+        return $this->hasMany(AlternateConnection::class);
+    }
+
     public function company(): BelongsTo
     {
         return $this->belongsTo(Company::class);